+++ apps/ng-sa-shopping/src/app/filters/filter.service.ts
@@ -1,9 +1,5 @@
-import { Injectable } from '@angular/core';
-import { adapt } from '@state-adapt/angular';
+import { adaptInjectable } from '@state-adapt/angular';
 import { filters } from '../../../../../libs/shopping/src';
 import { filterAdapter } from './filter.adapter';
 
-@Injectable({ providedIn: 'root' })
-export class FilterService {
-  filterStore = adapt(['filters', filters], filterAdapter);
-}
+export const injectFilterStore = adaptInjectable(['filters', filters], filterAdapter);