+++ apps/ng-sa-shopping/src/app/products/product.service.ts
@@ -1,35 +1,41 @@
 import { Injectable } from '@angular/core';
 import { joinStores, Source } from '@state-adapt/rxjs';
-import { adapt } from '@state-adapt/angular';
+import { adaptInjectable } from '@state-adapt/angular';
 import { Product, products, QuantityChange } from '../../../../../libs/shopping/src';
-import { FilterService } from '../filters/filter.service';
+import { injectFilterStore } from '../filters/filter.service';
 import { getFilteredProducts } from './filter-product.selectors';
 import { productAdapter } from './product.adapter';
 
+export const productSources = {
+  quantityChange$: new Source<QuantityChange>('[Products] quantityChange$'),
+  addToCart$: new Source<Product>('[Products] addToCart$'),
+  removeFromCart$: new Source<Product>('[Products] removeFromCart$'),
+};
+
+export const injectProductStore = adaptInjectable(
+  ['products', products, productAdapter],
+  {
+    changeQuantity: productSources.quantityChange$,
+    addProduct: productSources.removeFromCart$,
+    removeProduct: productSources.addToCart$,
+  },
+);
+
+export const injectCartStore = adaptInjectable(
+  ['cart', [] as Product[], productAdapter],
+  {
+    changeQuantity: productSources.quantityChange$,
+    addProduct: productSources.addToCart$,
+    removeProduct: productSources.removeFromCart$,
+  },
+);
+
 @Injectable({ providedIn: 'root' })
 export class ProductService {
-  quantityChange$ = new Source<QuantityChange>('[Products] quantityChange$');
-  addToCart$ = new Source<Product>('[Products] addToCart$');
-  removeFromCart$ = new Source<Product>('[Products] removeFromCart$');
-
-  productStore = adapt(['products', products, productAdapter], {
-    changeQuantity: this.quantityChange$,
-    addProduct: this.removeFromCart$,
-    removeProduct: this.addToCart$,
-  });
-
-  cartStore = adapt(['cart', [] as Product[], productAdapter], {
-    changeQuantity: this.quantityChange$,
-    addProduct: this.addToCart$,
-    removeProduct: this.removeFromCart$,
-  });
-
   filteredProductStore = joinStores({
-    product: this.productStore,
-    filter: this.filterService.filterStore,
+    product: injectProductStore(),
+    filter: injectFilterStore(),
   })({
     filteredProducts: s => getFilteredProducts(s.filter, s.product),
   })();
-
-  constructor(private filterService: FilterService) {}
 }