+++ libs/angular/src/index.ts
@@ -2,5 +2,6 @@ export * from './lib/provide-store.function';
 export * from './lib/default-store-provider.const';
 export * from './lib/adapt.function';
 export * from './lib/watch.function';
+export * from './lib/adapt-injectable.function';
 
 export * from './cheat-sheet.const';